feat: Avatar.Group (#25192)
