+++ components/avatar/style/index.less
@@ -61,3 +61,6 @@
     }
   }
 }
+
+@import './group';
+@import './rtl';
