+++ components/avatar/style/index.tsx
@@ -1,2 +1,5 @@
 import '../../style/index.less';
 import './index.less';
+
+// style dependencies
+import '../../popover/style';
