+++ components/avatar/style/rtl.less
@@ -1,0 +1,20 @@
+@import '../../style/themes/index';
+@import '../../style/mixins/index';
+
+@avatar-prefix-cls: ~'@{ant-prefix}-avatar';
+
+.@{avatar-prefix-cls}-group {
+  &-rtl {
+    .@{avatar-prefix-cls}:not(:first-child) {
+      margin-right: @avatar-group-overlapping;
+      margin-left: 0;
+    }
+  }
+
+  &-popover.@{ant-prefix}-popover-rtl {
+    .@{ant-prefix}-avatar + .@{ant-prefix}-avatar {
+      margin-right: @avatar-group-space;
+      margin-left: 0;
+    }
+  }
+}
